% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse}
\alias{browse}
\title{Browse QTL results}
\usage{
browse(scan1output, map, dir = NULL)
}
\arguments{
\item{scan1output}{Genome scan output, as from \code{\link[qtl2:scan1]{qtl2::scan1()}}}

\item{map}{Corresponding physical map, as a list of chromosomes
that are each a vector of marker positions.}

\item{dir}{Optional directory to contain the results. If not
provided, a temporary directory is created.}
}
\value{
File location (hidden).
}
\description{
Use the genetics genome browser to browse QTL mapping results
}
\examples{
# run qtl2
# browse results
}
