% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl2browse-package.R
\docType{package}
\name{qtl2browse-package}
\alias{qtl2browse}
\alias{qtl2browse-package}
\title{qtl2browse: Genetics Genome Browser}
\description{
Plot genome scan results with the Genetics Genome Browser: fast, lightweight, powered by purescript.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rqtl/qtl2browse}
  \item \url{https://kbroman.org/qtl2}
  \item Report bugs at \url{https://github.com/rqtl/qtl2browse/issues}
}

}
\author{
\strong{Maintainer}: Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})

Authors:
\itemize{
  \item Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})
  \item Christian Fischer
}

}
\keyword{internal}
