% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse}
\alias{browse}
\title{Browse QTL results}
\usage{
browse(scan1output, map, lodcolumn = 1, dir = NULL)
}
\arguments{
\item{scan1output}{Genome scan output, as from \code{\link[qtl2:scan1]{qtl2::scan1()}}}

\item{map}{Corresponding physical map, as a list of chromosomes
that are each a vector of marker positions.}

\item{lodcolumn}{LOD score column to plot (a numeric index, or a
character string for a column name). Only one value allowed.}

\item{dir}{Optional directory to contain the results. If not
provided, a temporary directory is created.}
}
\value{
File location (hidden).
}
\description{
Use the genetics genome browser to browse QTL mapping results
}
\examples{
\donttest{
library(qtl2)
recla <- read_cross2(paste0("https://raw.githubusercontent.com/rqtl/",
                            "qtl2data/master/DO_Recla/recla.zip"))

gmap <- insert_pseudomarkers(recla$gmap, step=0.2, stepwidth="max")
pmap <- interp_map(gmap, recla$gmap, recla$pmap)
pr <- calc_genoprob(recla, gmap, error_prob=0.002,
                    map_function="c-f", cores=0)
apr <- genoprob_to_alleleprob(pr)

k <- calc_kinship(apr, "loco", cores=0)

out <- scan1(apr, recla$pheno[,"HP_latency"], k, cores=0)

library(qtl2browse)
browse(out, pmap)
}
}
